defaultEnviron to base class
