+++ test/common.py
@@ -59,6 +59,21 @@ def make_zip(filename: str, data: str, n: int) -> bytes:
 
 class EngineBaseTest(unittest.TestCase):
 
+    @staticmethod
+    def defaultEnviron():
+        os.environ['VTT_TITLE'] = 'unittest'
+        os.environ['VTT_LIMIT_TOKEN'] =' 2'
+        os.environ['VTT_LIMIT_BG'] = '10'
+        os.environ['VTT_LIMIT_GAME'] = '5'
+        os.environ['VTT_LIMIT_MUSIC'] = '10'
+        os.environ['VTT_NUM_MUSIC'] = '5'
+        os.environ['VTT_CLEANUP_EXPIRE'] = '3600'
+        os.environ['VTT_CLEANUP_TIME'] = '03:00'
+        os.environ['VTT_DOMAIN'] = 'vtt.example.com'
+        os.environ['VTT_PORT'] = '8080'
+        os.environ.pop('VTT_SSL', None)
+        os.environ.pop('VTT_REVERSE_PROXY', None)
+
     def setUp(self) -> None:
         # create temporary directory
         self.tmpdir = tempfile.TemporaryDirectory()
@@ -76,6 +91,7 @@ class EngineBaseTest(unittest.TestCase):
         self.app = webtest.TestApp(self.engine.app)
         
         self.monkeyPatch()
+        EngineBaseTest.defaultEnviron()
                 
     def monkeyPatch(self) -> None:
         # save methods for later
