+++ vtt/routes/__init__.py
@@ -2,3 +2,4 @@ from .gm import register as register_gm
 from .player import register as register_player
 from .resources import register as register_resources
 from .error import register as register_error
+from .json_api import register as register_api
